refactor(plugins): continue overhaul -- more plugins converted to new format
