--- lua/plugins/coding/comment.lua
@@ -1,13 +1,0 @@
-return {
-  'numToStr/Comment.nvim',
-  event = 'VeryLazy',
-  -- event = { 'BufReadPost', 'BufNewFile' },
-  dependencies = {},
-  config = function()
-    -- import comment plugin safely
-    local comment = require('Comment')
-
-    -- enable comment
-    comment.setup({})
-  end,
-}
