+++ lua/plugins/disabled.lua
@@ -4,7 +4,6 @@
 
 return {
   { 'shellRaining/hlchunk.nvim', enabled = true },
-  { 'mcauley-penney/tidy.nvim', enabled = true },
   { 'Bekaboo/deadcolumn.nvim', enabled = true },
   { 'm4xshen/smartcolumn.nvim', enabled = false },
   { 'goolord/alpha-nvim', enabled = false },
