--- lua/plugins/editor/tidy.lua
@@ -1,9 +1,0 @@
-return {
-  {
-    'mcauley-penney/tidy.nvim',
-    event = 'LazyFile',
-    opts = {
-      filetype_exclude = { 'diff', 'gitcommit', 'markdown' },
-    },
-  },
-}
