--- lua/plugins/ui/notify.lua
@@ -1,32 +1,0 @@
-return {
-  {
-    'rcarriga/nvim-notify',
-    keys = {
-      {
-        '<leader>un',
-        function()
-          require('notify').dismiss({ silent = true, pending = true })
-        end,
-        desc = 'Dismiss all notifications',
-      },
-    },
-    opts = {
-      timeout = 3000,
-      max_height = function()
-        return math.floor(vim.o.lines * 0.75)
-      end,
-      max_width = function()
-        return math.floor(vim.o.columns * 0.75)
-      end,
-    },
-    init = function()
-      -- If noice is not enabled, Load  notify on VeryLazy
-      local util = require('core.util')
-      if not util.has('noice.nvim') then
-        util.on_very_lazy(function()
-          vim.notify = require('notify')
-        end)
-      end
-    end,
-  },
-}
