--- lua/plugins/utils/wakatime.lua
@@ -1,6 +1,0 @@
-return {
-  {
-    'wakatime/vim-wakatime',
-    event = 'InsertEnter',
-  },
-}
